+++ Products/Poi/browser/response.py
@@ -436,7 +436,7 @@ class Create(Base):
                 new_label = self.available_releases.getValue(new)
                 current_label = self.available_releases.getValue(current)
                 changes['targetRelease'] = new
-                new_response.add_change(option, _(u'Target release'),
+                new_response.add_change('targetRelease', _(u'Target release'),
                                         current_label, new_label)
                 issue_has_changed = True